api/requirements.txt
# Django packages
django-braces==1.13.0
django-cors-headers==2.4.0
django-extensions==2.1.3
django-filter==2.0.0
django-countries==5.3.2

# REST
drf-amsterdam==0.1.7
drf-extensions==0.4.0
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
djangorestframework-xml==1.3.0
djangorestframework-gis==0.13
drf-yasg==1.10.2
# Database
psycopg2-binary==2.7.5
# Datapunt packages
datapunt-authorization-django==0.2.23
datapunt-authorization-levels==0.1.3
# Validation
swagger-spec-validator==2.4.1
# Logging
graypy==0.3.1
pygelf==0.3.4
# Markup
# Sentry
raven==6.9.0
## The following requirements were added by pip freeze:
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
defusedxml==0.5.0
Django==2.1.2
django-debug-toolbar==1.10.1
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
requests==2.19.1
ruamel.yaml==0.15.72
six==1.11.0
sqlparse==0.2.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
